API/requirements.txt
Flask
pyjwt
